abstracted fetchPlaces out froom component to redux action and displaying results from redux state instead of component state
